tests/test_trade_executor.py
"""
Unit tests for trade execution functionality.
"""
import pytest
import asyncio
from unittest.mock import Mock, patch
from datetime import datetime
from src.trading.trade_executor import TradeExecutor
from src.trading.dex_interface import RaydiumDEX
from src.wallet.wallet_manager import WalletManager
from src.utils.config import Config

@pytest.fixture
def config():
    """Create Config instance for testing"""
    return Config()

@pytest.fixture
async def trade_executor(config):
    """Create TradeExecutor instance for testing"""
    wallet_manager = Mock(spec=WalletManager)
    dex = Mock(spec=RaydiumDEX)
    executor = TradeExecutor(wallet_manager, dex, config.to_dict())
    return executor

@pytest.mark.asyncio
async def test_execute_trading_round(trade_executor):
    """Test trading round execution"""
    # Mock wallet creation
    trade_executor.wallet_manager.create_wallet_group.return_value = [
        f'wallet{i}' for i in range(5)
    ]

    # Mock SOL distribution
    trade_executor.wallet_manager.distribute_sol.return_value = True

    # Execute trading round
    await trade_executor.execute_trading_round()

    # Verify wallet operations
    trade_executor.wallet_manager.create_wallet_group.assert_called_once()
    trade_executor.wallet_manager.distribute_sol.assert_called_once()
    trade_executor.wallet_manager.recall_sol.assert_called_once()

@pytest.mark.asyncio
async def test_high_frequency_trading(trade_executor):
    """Test high-frequency trading execution"""
    # Enable high-frequency trading
    trade_executor.high_frequency['enabled'] = True

    # Mock wallet operations
    trade_executor.wallet_manager.create_wallet_group.return_value = [
        f'wallet{i}' for i in range(10)
    ]

    # Execute high-frequency round
    await trade_executor.execute_high_frequency_round()

    # Verify trading operations
    assert trade_executor.wallet_manager.create_wallet_group.called
    assert trade_executor.trading_pattern.execute_pump_pattern.called or \
           trade_executor.trading_pattern.execute_milkshake_pattern.called

@pytest.mark.asyncio
async def test_trading_session(trade_executor):
    """Test complete trading session"""
    # Mock time to control session duration
    with patch('datetime.datetime') as mock_datetime:
        mock_datetime.now.side_effect = [
            datetime(2024, 1, 1, 12, 0),  # Start time
            datetime(2024, 1, 1, 12, 30),  # Mid-session check
            datetime(2024, 1, 1, 13, 0),  # End time
        ]

        # Start trading session
        await trade_executor.start_trading_session()

        # Verify session execution
        assert trade_executor.wallet_manager.create_wallet_group.called
        assert trade_executor.trading_pattern.execute_pump_pattern.called or \
               trade_executor.trading_pattern.execute_milkshake_pattern.called

@pytest.mark.asyncio
async def test_error_handling(trade_executor):
    """Test error handling in trade execution"""
    # Mock wallet creation failure
    trade_executor.wallet_manager.create_wallet_group.side_effect = Exception("Creation failed")

    # Verify error handling
    with pytest.raises(Exception):
        await trade_executor.execute_trading_round()
